Fix APR calculation for v1 crypto pools
