+++ packages/utils/pricing.ts
@@ -2,12 +2,14 @@ import { Address, BigDecimal, BigInt, log } from '@graphprotocol/graph-ts'
 import {
   ADDRESS_ZERO,
   BIG_DECIMAL_1E18,
+  BIG_DECIMAL_1E6,
   BIG_DECIMAL_ONE,
   BIG_DECIMAL_ZERO,
   BIG_INT_ZERO,
   CRV_FRAX_ADDRESS,
   CTOKENS,
   FRAXBP_ADDRESS,
+  RKP3R_ADDRESS,
   SUSHI_FACTORY_ADDRESS,
   THREE_CRV_ADDRESS,
   TRIPOOL_ADDRESS,
@@ -28,6 +30,7 @@ import { exponentToBigDecimal, exponentToBigInt } from './maths'
 import { FactoryV3 } from 'convex/generated/Booster/FactoryV3'
 import { Quoter } from 'convex/generated/Booster/Quoter'
 import { CurvePool } from 'convex/generated/Booster/CurvePool'
+import { RedeemableKeep3r } from 'convex/generated/Booster/RedeemableKeep3r'
 
 export function getEthRate(token: Address): BigDecimal {
   let eth = BIG_DECIMAL_ONE
@@ -167,10 +170,21 @@ export function get3CrvVirtualPrice(): BigDecimal {
   return vPrice
 }
 
+function getRKp3rPrice(): BigDecimal {
+  const RKp3rContract = RedeemableKeep3r.bind(RKP3R_ADDRESS)
+  const discount = RKp3rContract.discount()
+  const priceResult = RKp3rContract.try_price()
+  if (priceResult.reverted) {
+    return BIG_DECIMAL_ZERO
+  }
+  return priceResult.value.times(discount).div(BigInt.fromI32(100)).toBigDecimal().div(BIG_DECIMAL_1E6)
+}
+
 export function getUsdRate(token: Address): BigDecimal {
   const usdt = BIG_DECIMAL_ONE
-
-  if (token == CRV_FRAX_ADDRESS) {
+  if (token == RKP3R_ADDRESS) {
+    return getRKp3rPrice()
+  } else if (token == CRV_FRAX_ADDRESS) {
     return getFraxBpVirtualPrice()
   } else if (token == THREE_CRV_ADDRESS) {
     return get3CrvVirtualPrice()
